+++ MANIFEST.in
@@ -1,2 +1,2 @@
-include src/west/commands/west-commands-schema.yml
+include src/west/west-commands-schema.yml
 include src/west/manifest-schema.yml
