+++ src/west/app/README.txt
@@ -1,0 +1,8 @@
+The west.app package contains the implementation of the west command
+line tool and its built-in features.
+
+Nothing in this package should be considered API.
+
+In particular, authors of west extension commands should not rely on
+the behavior this package or any of its contents to remain stable over
+time.
