+++ src/west/app/__init__.py
@@ -1,0 +1,3 @@
+# Copyright (c) 2020, Nordic Semiconductor ASA
+
+# nothing here
