+++ src/west/west-commands-schema.yml
